+++ battle.py
@@ -1,4 +1,6 @@
 from classes import *
+from characters import *
+from items import *
 from infos import *
 from decorators import *
