+++ temp.py
@@ -1,223 +1,12 @@
-from infos import *
-from classes import *
+from characters import *
+from world import Area
 from battle import Battle
-from world import *
-import time
+from objects import *
+from decorators import print_one_line_in_frame
 from collections import OrderedDict
+import os
+import time
 
-
-# Create Hero and Enemy
-player = Hero(
-    "Thorin Oakenshield",
-    1,
-    0,
-    "Dwarf",
-    110,
-    94,
-    20,
-    8,
-    15,
-    "Sick",
-    None,
-    True,
-    20,
-    10,
-)
-
-enemy = Enemy(
-    "Azog The Defiler",
-    1,
-    25,
-    "Orc",
-    70,
-    70,
-    10,
-    5,
-    10,
-    "Stunned",
-    None,
-    True,
-    "Cold",
-    "Fire",
-)
-
-# Create weapons
-excalibur = Weapon(
-    "Excalibur",
-    "Great Sword",
-    50,
-    1,
-    15,
-    15,
-    "Human",
-    5,
-    2,
-    Inventory,
-    "Slash",
-    15,
-    False,
-)
-
-iceblizzard = Weapon(
-    "Ice Blizzard",
-    "Staff",
-    40,
-    1,
-    10,
-    15,
-    "Human",
-    10,
-    10,
-    Inventory,
-    "ice",
-    10,
-    False,
-)
-
-thorshammer = Weapon(
-    "Destroyer",
-    "Hammer",
-    10,
-    1,
-    20,
-    10,
-    "Dwarf",
-    20,
-    20,
-    Inventory,
-    "blunt",
-    20,
-    False,
-)
-
-elvisheyes = Weapon(
-    "Elvish Eyes",
-    "Bow",
-    20,
-    1,
-    15,
-    15,
-    "Elf",
-    25,
-    25,
-    Inventory,
-    "stab",
-    12,
-    False,
-)
-
-longspear = Weapon(
-    "Long Spear",
-    "Spear",
-    15,
-    2,
-    10,
-    15,
-    "Human",
-    10,
-    10,
-    Inventory,
-    "stab",
-    15,
-    False,
-)
-
-silver_plate = Armor(
-    "Silver Plate",
-    "Great Armor made of silver.",
-    30,
-    2,
-    20,
-    0,
-    "Human",
-    40,
-    40,
-    Inventory,
-    "Cold",
-    "Blunt",
-    False,
-)
-
-leather_armor = Armor(
-    name="Leather Jacket",
-    description="Good for Humans",
-    value=30,
-    slot_size=1,
-    required_strength=5,
-    required_dexterity=0,
-    allowed_race="Human",
-    max_durability=15,
-    durability=10,
-    inventory=Inventory,
-    resistance="Cold",
-    protection=5,
-    is_equipped=False,
-)
-
-life_potion = Consumable(
-    name="Life Potion",
-    description="Restores 20 health.",
-    value=20,
-    slot_size=1,
-    required_strength=0,
-    required_dexterity=0,
-    allowed_race="Human",
-    max_durability=1,
-    durability=1,
-    inventory=Inventory,
-    heal=20,
-    mana=0,
-    strength=0,
-    dexterity=0,
-)
-
-boost_potion = Consumable(
-    name="Boost Potion",
-    description="Restores 10 health and 10 mana.",
-    value=20,
-    slot_size=1,
-    required_strength=0,
-    required_dexterity=0,
-    allowed_race="Human",
-    max_durability=1,
-    durability=1,
-    inventory=Inventory,
-    heal=90,
-    mana=10,
-    strength=0,
-    dexterity=0,
-)
-
-strength_potion = Consumable(
-    name="Strenght Potion",
-    description="Increases Strength by 20",
-    value=20,
-    slot_size=1,
-    required_strength=0,
-    required_dexterity=0,
-    allowed_race="Human",
-    max_durability=1,
-    durability=1,
-    inventory=Inventory,
-    heal=0,
-    mana=0,
-    strength=15,
-    dexterity=0,
-)
-
-# Create inventory and add some items
-player.inventory = Inventory([], 20, 20)
-enemy.inventory = Inventory([], 5, 5)
-player.inventory.add_item(excalibur)
-player.inventory.add_item(iceblizzard)
-player.inventory.add_item(thorshammer)
-player.inventory.add_item(elvisheyes)
-player.inventory.add_item(longspear)
-enemy.inventory.add_item(silver_plate)
-player.inventory.add_item(life_potion)
-player.inventory.add_item(boost_potion)
-player.inventory.add_item(strength_potion)
-player.inventory.add_item(leather_armor)
 
 ### Health Bar creation and drawing
 # player_healthbar = HealthBar(player)
@@ -268,12 +57,12 @@ town = Area(
     visited=False,
 )
 
-areas = {
-    "Forest": forest,
-    "Town": town,
-    "Plains": None,
-    "Ruins": None,
-}
+# areas = {
+#     "Forest": forest,
+#     "Town": town,
+#     "Plains": None,
+#     "Ruins": None,
+# }
 
 
 def explore_area(area):
